+++ hpx/include/threadmanager.hpp
@@ -11,7 +11,7 @@
 
 #include <hpx/runtime/thread_pool_helpers.hpp>
 #include <hpx/runtime/threads/executors/embedded_thread_pool_executors.hpp>
-#include <hpx/thread.hpp>
+#include <hpx/threading.hpp>
 #include <hpx/threading_base/thread_data.hpp>
 #include <hpx/threading_base/thread_helpers.hpp>
 #include <hpx/runtime/threads/threadmanager.hpp>
