+++ hpx/lcos/detail/async_implementations.hpp
@@ -15,7 +15,7 @@
 #include <hpx/runtime/launch_policy.hpp>
 #include <hpx/runtime/naming/address.hpp>
 #include <hpx/runtime/naming/id_type.hpp>
-#include <hpx/thread.hpp>
+#include <hpx/threading.hpp>
 #include <hpx/threading_base/thread_init_data.hpp>
 #include <hpx/errors.hpp>
 #include <hpx/traits/action_decorate_function.hpp>
