+++ libs/CMakeLists.txt
@@ -59,7 +59,7 @@ set(HPX_LIBS
   string_util
   synchronization
   testing
-  thread
+  threading
   thread_pools
   thread_support
   threading_base
