+++ libs/all_modules.rst
@@ -62,7 +62,7 @@ All modules
    /libs/string_util/docs/index.rst
    /libs/synchronization/docs/index.rst
    /libs/testing/docs/index.rst
-   /libs/thread/docs/index.rst
+   /libs/threading/docs/index.rst
    /libs/thread_pools/docs/index.rst
    /libs/thread_support/docs/index.rst
    /libs/threading_base/docs/index.rst
