+++ libs/collectives/CMakeLists.txt
@@ -78,7 +78,7 @@ add_hpx_module(collectives
       hpx_serialization
       hpx_synchronization
       hpx_timing
-      hpx_thread
+      hpx_threading
       hpx_thread_support
       hpx_threading_base
       hpx_threadmanager
