+++ libs/synchronization/tests/unit/condition_variable.cpp
@@ -12,7 +12,7 @@
 #include <hpx/hpx_init.hpp>
 #include <hpx/synchronization/condition_variable.hpp>
 #include <hpx/synchronization/mutex.hpp>
-#include <hpx/thread.hpp>
+#include <hpx/threading.hpp>
 #include <hpx/topology/topology.hpp>
 
 #include <hpx/testing.hpp>
