+++ libs/synchronization/tests/unit/shared_mutex/thread_group.hpp
@@ -13,7 +13,7 @@
 #include <hpx/errors/exception.hpp>
 #include <hpx/synchronization/mutex.hpp>
 #include <hpx/synchronization/shared_mutex.hpp>
-#include <hpx/thread.hpp>
+#include <hpx/threading.hpp>
 
 #include <boost/thread/locks.hpp>
