+++ libs/threading/docs/index.rst
@@ -5,11 +5,11 @@
     Distributed under the Boost Software License, Version 1.0. (See accompanying
     file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-.. _libs_thread:
+.. _libs_threading:
 
-======
-thread
-======
+=========
+threading
+=========
 
 TODO: High-level description of the library.
