+++ libs/threading/include_compatibility/hpx/runtime/threads/thread.hpp
@@ -5,15 +5,15 @@
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 #include <hpx/config.hpp>
-#include <hpx/thread/config/defines.hpp>
-#include <hpx/thread/thread.hpp>
+#include <hpx/threading/config/defines.hpp>
+#include <hpx/threading/thread.hpp>
 
-#if defined(HPX_THREAD_HAVE_DEPRECATION_WARNINGS)
+#if defined(HPX_THREADING_HAVE_DEPRECATION_WARNINGS)
 #if defined(HPX_MSVC)
 #pragma message("The header hpx/runtime/threads/thread.hpp is deprecated, \
-    please include hpx/thread/thread.hpp instead")
+    please include hpx/threading/thread.hpp instead")
 #else
 #warning "The header hpx/runtime/threads/thread.hpp is deprecated, \
-    please include hpx/thread/thread.hpp instead"
+    please include hpx/threading/thread.hpp instead"
 #endif
 #endif
