+++ libs/threading/src/thread.cpp
@@ -13,8 +13,8 @@
 #include <hpx/lcos/detail/future_data.hpp>
 #include <hpx/lcos/future.hpp>
 #include <hpx/memory/intrusive_ptr.hpp>
-#include <hpx/thread.hpp>
 #include <hpx/thread_support/unlock_guard.hpp>
+#include <hpx/threading.hpp>
 #include <hpx/threading_base/thread_helpers.hpp>
 #include <hpx/threading_base/thread_init_data.hpp>
 #include <hpx/threading_base/thread_pool_base.hpp>
