+++ libs/threading/tests/CMakeLists.txt
@@ -8,34 +8,34 @@ include(HPX_Message)
 include(HPX_Option)
 
 if (NOT HPX_WITH_TESTS AND HPX_TOP_LEVEL)
-  hpx_set_option(HPX_THREAD_WITH_TESTS VALUE OFF FORCE)
+  hpx_set_option(HPX_THREADING_WITH_TESTS VALUE OFF FORCE)
   return()
 endif()
 
-if (HPX_THREAD_WITH_TESTS)
+if (HPX_THREADING_WITH_TESTS)
     if (HPX_WITH_TESTS_UNIT)
-      add_hpx_pseudo_target(tests.unit.modules.thread)
-      add_hpx_pseudo_dependencies(tests.unit.modules tests.unit.modules.thread)
+      add_hpx_pseudo_target(tests.unit.modules.threading)
+      add_hpx_pseudo_dependencies(tests.unit.modules tests.unit.modules.threading)
       add_subdirectory(unit)
     endif()
 
     if (HPX_WITH_TESTS_REGRESSIONS)
-      add_hpx_pseudo_target(tests.regressions.modules.thread)
-      add_hpx_pseudo_dependencies(tests.regressions.modules tests.regressions.modules.thread)
+      add_hpx_pseudo_target(tests.regressions.modules.threading)
+      add_hpx_pseudo_dependencies(tests.regressions.modules tests.regressions.modules.threading)
       add_subdirectory(regressions)
     endif()
 
     if (HPX_WITH_TESTS_BENCHMARKS)
-      add_hpx_pseudo_target(tests.performance.modules.thread)
-      add_hpx_pseudo_dependencies(tests.performance.modules tests.performance.modules.thread)
+      add_hpx_pseudo_target(tests.performance.modules.threading)
+      add_hpx_pseudo_dependencies(tests.performance.modules tests.performance.modules.threading)
       add_subdirectory(performance)
     endif()
 
     if (HPX_WITH_TESTS_HEADERS)
       add_hpx_header_tests(
-        modules.thread
-        HEADERS ${thread_headers}
+        modules.threading
+        HEADERS ${threading_headers}
         HEADER_ROOT ${PROJECT_SOURCE_DIR}/include
-        DEPENDENCIES hpx_thread)
+        DEPENDENCIES hpx_threading)
     endif()
 endif()
