+++ plugins/parcelport/tcp/CMakeLists.txt
@@ -35,7 +35,7 @@ if(HPX_WITH_PARCELPORT_TCP)
       hpx_plugin
       hpx_program_options
       hpx_serialization
-      hpx_thread
+      hpx_threading
       hpx_threadmanager
       hpx_timing
       hpx_topology
